Extract Distribution families of the gamlss.dist Package

The functions searches in the gamlss.dist namespace for available distributions. It returns a named list of faily functions which can be used with bamlss.

gamlss_distributions(type = c("continuous", "discrete"))

Arguments

type

Character specifying the type of distribution to be extracted.

See also

Examples

if (FALSE) {
dists <- gamlss_distributions(type = "continuous")
print(dists)
}