Kriging Smooth Constructor

This smooth constructor implements a kriging based model term.

# S3 method for kr.smooth.spec
smooth.construct(object, data, knots, ...)
# S3 method for kriging.smooth
Predict.matrix(object, data)

Arguments

object, data, knots

See smooth.construct.

...

Currently not used.

Details

This smooth constructor implements univariate and bivariate Kriging terms. The basis functions are based on the Matern covariance function. For finding knots, a space filling algorithm is used, see cover.design.

Value

A smooth specification object, see also smooth.construct.

References

Fahrmeir, L., Kneib, T. , Lang, S., Marx, B. (2013): Regression. Models, Methods and Applications, Springer Verlag. https://www.uni-goettingen.de/de/551357.html

See also

Examples

if (FALSE) ## Simulate data.
set.seed(123)
d <- GAMart()

## Estimate model.
f <- num ~ s(x1,bs="kr") + s(x2,bs="kr") + s(x3,bs="kr") + s(lon,lat,bs="kr",k=30)

## Set the seed, estimate model.
set.seed(111)
b <- bamlss(f, data = d)
#> Error in UseMethod("smooth.construct"): no applicable method for 'smooth.construct' applied to an object of class "kr.smooth.spec"

## Plot estimated effects.
plot(b)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': object 'b' not found