Summary for BAMLSS

The function takes an object of class "bamlss" and produces summaries of optimizer and sampler function outputs.

# S3 method for bamlss
summary(object, model = NULL,
  FUN = NULL, parameters = TRUE, ...)

# S3 method for summary.bamlss
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object
An object of class "bamlss".

x
An oject of class "summary.bamlss".

model
Character or integer, specifies the model for which a summary should be computed.

FUN
Function that should be applied on samples, see also function coef.bamlss.

parameters
If an optimizer function is applied within the bamlss call, should the values of the estimated parameters be part of the summary?

digits
Controls number of digits printed in output.

…
Other arguments.

Details

If the fitted model contains samples, summaries according to the supplied function can be computed, e.g., different quantiles of samples. See also function coef.bamlss that extracts the coefficient summaries.

If an optimizer function was used within the bamlss call, estimated parameters will be included per default into the summary.

Note that summaries not based on samples can be user defined, e.g., as returned from function samplestats or the return values of optimizer function, e.g., see function bfit.

Value

summary.bamlss produces the following summary:

call

The initial bamlss call.

family

The family that is used for modeling.

formula

The model formula.

model.matrix

Summary of parameteric terms.

smooth.construct

Summary of smooth terms.

model.stats

Other model statistics, e.g., as returned from optimizer functions and/or produces by function samplestats.

See also

Examples

# NOT RUN {
## Generate some data.
d <- GAMart()

## Model formula.
f <- list(
  num ~ s(x1) + s(x2),
  sigma ~ s(x3) + te(lon,lat)
)

## Estimate model.
b <- bamlss(f, data = d)

## Print the summary.
print(summary(b))
# }