Batchwise Backfitting

Batchwise backfitting estimation engine for GAMLSS using very large data sets.

## Batchwise backfitting engine.
opt_bbfit(x, y, family, shuffle = TRUE, start = NULL, offset = NULL,
  epochs = 1, nbatch = 10, verbose = TRUE, ...)

bbfit(x, y, family, shuffle = TRUE, start = NULL, offset = NULL,
  epochs = 1, nbatch = 10, verbose = TRUE, ...)

## Parallel version.
opt_bbfitp(x, y, family, mc.cores = 1, ...)

## Loglik contribution plot.
contribplot(x, ...)

Arguments

x

For function bfit() the x list, as returned from function bamlss.frame, holding all model matrices and other information that is used for fitting the model. For the updating functions an object as returned from function smooth.construct or smoothCon. For function contribplot(), a "bamlss" object using bbfit() with argument select = TRUE.

y

The model response, as returned from function bamlss.frame.

family

A bamlss family object, see family.bamlss.

shuffle

Should observations be shuffled?

start

A named numeric vector containing possible starting values, the names are based on function parameters.

offset

Can be used to supply model offsets for use in fitting, returned from function bamlss.frame.

epochs

For how many epochs should the algorithm run?

nbatch

Number of batches. Can also be a number between 0 and 1, i.e., determining the fraction of observations that should be used for fitting.

verbose

Print information during runtime of the algorithm.

mc.cores

On how many cores should estimation be started?

...

For bbfitp() all arguments to be passed to bbfit().

Details

The algorithm uses batch-wise estimation of regression coefficients and smoothing variances. The smoothing variances are estimated on an hold-out batch. This way, models for very large data sets can be estimated. Note, the algorithm can work in combination with the ff and ffbase package, i.e., the entire data is never in the computer RAM. Therefore, the data can either to be stored as comma separated file on disc or provided as "ffdf" data frame, see also the examples.

The optimizer functions use additional arguments:

batch_ids. This argument can either be a list of indices specifying the batches that should be used for estimation, or a vector of length 2, where the first element specifies the number of observations that should be sampled for each batch and the second argument specifies the number of batches, see the example.

nu, the step length control parameter. Defaults to nu = 0.05. If argument slice = TRUE then nu = 1.

loglik, defaults to loglik = FALSE. If set to loglik = TRUE the "out-of-sample" log-likelihood is used for smoothing variance estimation.

aic, defaults to aic = FALSE, If set to aic = TRUE the "out-of-sample" AIC is used for smoothing variance estimation.

eps_loglik, defaults to eps_loglik = 0.01. This argument specifies the relative change in the "out-of-sample" log-likelihood that is needed such that a model term gets updated.

select, defaults to select = FALSE. If set to select = TRUE, the algorithm only selects the model term with the largest contribution in the "out-of-sample" log-likelihood for updating in each iteration/batch.

always, defaults to always = FALSE. If set to always = TRUE no log-likelihood contribution checks will be used and model terms are always updated.

K, defaults to K = 2. This argument controls the penalty on the degrees of freedom in the computation of the AIC.

slice, defaults to slice = FALSE. If set to slice = TRUE, slice sampling using the "out-of-sample" log-likelihood or AIC is used for smoothing variance estimation. Moreover, always = TRUE, eps_loglik = -Inf and nu = 1. If slice is an integer n, slice sampling is started after n iterations, before smoothing variances are optimized.

When using function opt_bbfitp, the parameter updates are stored as "mcmc" objects. In this case the traceplots can be visualized using plot.bamlss.

Value

For function opt_bbfit() a list containing the following objects:

fitted.values

A named list of the fitted values of the modeled parameters of the selected distribution.

parameters

The estimated set regression coefficients and smoothing variances.

shuffle

Logical

runtime

The runtime of the algorithm.

See also

Examples

if (FALSE) ## Simulate data.
set.seed(123)
d <- GAMart(n = 27000, sd = -1)

## Write data to disc.
tf <- tempdir()
write.table(d, file.path(tf, "d.raw"), quote = FALSE, row.names = FALSE, sep = ",")

## Model formula.
f <- list(
  y ~ s(x1,k=40) + s(x2,k=40) + s(x3,k=40) + te(lon,lat,k=10),
  sigma ~ s(x1,k=40) + s(x2,k=40) + s(x3,k=40) + te(lon,lat,k=10)
)

## Specify 50 batches with 1000 observations.
batch_ids <- c("nobs" = 1000, "nbatch" = 50)

## Note, can also be a list of indices, e.g.
## batch_ids <- lapply(1:50, function(i) { sample(1:nrow(d), size = 1000) })

## Different flavors:
## (1) Using "out-of-sample" aic for smoothing
##     variance estimation. Update is only accepted
##     if the "out-of-sample" log-likelihood is
##     increased. If data is a filepath, the data set is
##     read into R using package ff and model and
##     design matrices are processed with ff. This may
##     take some time depending on the size of the data.
set.seed(1)
b1 <- bamlss(f, data = file.path(tf, "d.raw"),
  sampler = FALSE, optimizer = opt_bbfit,
  batch_ids = batch_ids, nu = 0.1, aic = TRUE, eps_loglik = -Inf,
  always = FALSE)
#>   .. creating directory 'ff_data_bamlss' for storing matrices. Note, the directory is may not deleted and matrices can be used for another model. Use delete = TRUE in the bamlss call. Before starting a new model you can set overwrite = TRUE to overwrite existing data.
#>   .. ff processing model.matrix
#>   .. ..     100%
#>   .. ff processing term s(x1) 
#>   .. ..     100%
#>   .. ff processing term s(x2) 
#>   .. ..     100%
#>   .. ff processing term s(x3) 
#>   .. ..     100%
#>   .. ff processing term te(lon,lat) 
#>   .. ..     100%
#>   .. ff processing model.matrix
#>   .. ..     100%
#>   .. ff processing term s(x1) 
#>   .. ff processing term s(x2) 
#>   .. ff processing term s(x3) 
#>   .. ff processing term te(lon,lat) 
#>   .. df mu term s(x1) 4.158529 
#>   .. df mu term s(x2) 4.179289 
#>   .. df mu term s(x3) 4.174733 
#>   .. df mu term te(lon,lat) 10.00755 
#> opening ff ff_data_bamlss/sx1.ff
#>   .. df sigma term s(x1) 4.163124 
#> opening ff ff_data_bamlss/sx2.ff
#>   .. df sigma term s(x2) 4.180381 
#> opening ff ff_data_bamlss/sx3.ff
#>   .. df sigma term s(x3) 4.172028 
#> opening ff ff_data_bamlss/telonlat.ff
#>   .. df sigma term te(lon,lat) 10.00569 
#>    * iter 1, nobs 1000, edf 32.102600   * iter 2, nobs 2000, eps 1.618100, edf 38.820000   * iter 3, nobs 3000, eps 0.376100, edf 36.410000   * iter 4, nobs 4000, eps 0.264500, edf 27.890000   * iter 5, nobs 5000, eps 0.603800, edf 41.130000   * iter 6, nobs 6000, eps 0.328800, edf 32.070000   * iter 7, nobs 7000, eps 0.343000, edf 29.390000   * iter 8, nobs 8000, eps 0.196800, edf 27.370000   * iter 9, nobs 9000, eps 0.117100, edf 28.130000   * iter 10, nobs 10000, eps 0.116700, edf 32.000000   * iter 11, nobs 11000, eps 0.054400, edf 30.430000   * iter 12, nobs 12000, eps 0.103600, edf 33.160000   * iter 13, nobs 13000, eps 0.078300, edf 32.540000   * iter 14, nobs 14000, eps 0.104900, edf 27.680000   * iter 15, nobs 15000, eps 0.063900, edf 38.490000   * iter 16, nobs 16000, eps 0.039700, edf 34.150000   * iter 17, nobs 17000, eps 0.203100, edf 26.980000   * iter 18, nobs 18000, eps 0.058900, edf 30.440000   * iter 19, nobs 19000, eps 0.037900, edf 30.140000   * iter 20, nobs 20000, eps 0.054600, edf 34.070000   * iter 21, nobs 21000, eps 0.080500, edf 29.870000   * iter 22, nobs 22000, eps 0.036100, edf 32.810000   * iter 23, nobs 23000, eps 0.078900, edf 30.190000   * iter 24, nobs 24000, eps 0.019700, edf 21.430000   * iter 25, nobs 25000, eps 0.039300, edf 27.610000   * iter 26, nobs 26000, eps 0.034500, edf 22.560000   * iter 27, nobs 27000, eps 0.032200, edf 29.770000   * iter 28, nobs 28000, eps 0.044300, edf 32.190000   * iter 29, nobs 29000, eps 0.180300, edf 27.290000   * iter 30, nobs 30000, eps 0.089400, edf 36.050000   * iter 31, nobs 31000, eps 0.054500, edf 29.460000   * iter 32, nobs 32000, eps 0.023800, edf 27.410000   * iter 33, nobs 33000, eps 0.029300, edf 24.260000   * iter 34, nobs 34000, eps 0.033500, edf 27.210000   * iter 35, nobs 35000, eps 0.103000, edf 38.120000   * iter 36, nobs 36000, eps 0.061600, edf 32.070000   * iter 37, nobs 37000, eps 0.081700, edf 23.660000   * iter 38, nobs 38000, eps 0.033100, edf 19.940000   * iter 39, nobs 39000, eps 0.054600, edf 33.960000   * iter 40, nobs 40000, eps 0.038500, edf 19.320000   * iter 41, nobs 41000, eps 0.037800, edf 24.910000   * iter 42, nobs 42000, eps 0.060300, edf 28.370000   * iter 43, nobs 43000, eps 0.042200, edf 13.490000   * iter 44, nobs 44000, eps 0.109500, edf 22.420000   * iter 45, nobs 45000, eps 0.058300, edf 21.030000   * iter 46, nobs 46000, eps 0.042500, edf 20.030000   * iter 47, nobs 47000, eps 1.538200, edf 28.620000   * iter 48, nobs 48000, eps 0.033000, edf 20.970000   * iter 49, nobs 49000, eps 0.102600, edf 23.990000   * iter 50, nobs 50000, eps 0.000100, edf 0.000000
#> 
#> elapsed time:  2.29min

## Plot estimated effects.
## plot(b1)

## Plot coefficient paths for x3 in mu.
## pathplot(b1, name = "mu.s.s(x3).b")

## (2) Same but always update, this mimics the classic SGD.
##     Note, for prediction only the last iteration is
##     used in this case. To use more iterations use opt_bbfitp(),
##     Then iterations are stored as "mcmc" object and we can
##     predict using the burnin argment, e.g.,
##     p <- predict(b2, model = "mu", burnin = 35)
set.seed(2)
b2 <- bamlss(f, data = file.path(tf, "d.raw"),
  sampler = FALSE, optimizer = opt_bbfit,
  batch_ids = batch_ids, nu = 0.1, aic = TRUE, eps_loglik = -Inf,
  always = TRUE)
#>   .. creating directory 'ff_data_bamlss' for storing matrices. Note, the directory is may not deleted and matrices can be used for another model. Use delete = TRUE in the bamlss call. Before starting a new model you can set overwrite = TRUE to overwrite existing data.
#>   .. ff processing model.matrix
#>   .. ..     100%
#>   .. ff processing term s(x1) 
#>   .. ..     100%
#>   .. ff processing term s(x2) 
#>   .. ..     100%
#>   .. ff processing term s(x3) 
#>   .. ..     100%
#>   .. ff processing term te(lon,lat) 
#>   .. ..     100%
#>   .. ff processing model.matrix
#>   .. ..     100%
#>   .. ff processing term s(x1) 
#>   .. ff processing term s(x2) 
#>   .. ff processing term s(x3) 
#>   .. ff processing term te(lon,lat) 
#>   .. df mu term s(x1) 4.144054 
#>   .. df mu term s(x2) 4.167427 
#>   .. df mu term s(x3) 4.185937 
#>   .. df mu term te(lon,lat) 10.00854 
#> opening ff ff_data_bamlss/sx1.ff
#>   .. df sigma term s(x1) 4.178629 
#> opening ff ff_data_bamlss/sx2.ff
#>   .. df sigma term s(x2) 4.154461 
#> opening ff ff_data_bamlss/sx3.ff
#>   .. df sigma term s(x3) 4.165777 
#> opening ff ff_data_bamlss/telonlat.ff
#>   .. df sigma term te(lon,lat) 10.00654 
#>    * iter 1, nobs 1000, edf 50.250000   * iter 2, nobs 2000, eps 2.022300, edf 33.620000   * iter 3, nobs 3000, eps 0.428800, edf 39.590000   * iter 4, nobs 4000, eps 0.381000, edf 29.310000   * iter 5, nobs 5000, eps 0.301700, edf 35.810000   * iter 6, nobs 6000, eps 0.138000, edf 29.040000   * iter 7, nobs 7000, eps 0.112300, edf 34.540000   * iter 8, nobs 8000, eps 0.129400, edf 27.460000   * iter 9, nobs 9000, eps 0.097300, edf 27.700000   * iter 10, nobs 10000, eps 0.119900, edf 39.690000   * iter 11, nobs 11000, eps 0.091400, edf 34.830000   * iter 12, nobs 12000, eps 0.072700, edf 31.080000   * iter 13, nobs 13000, eps 0.066700, edf 30.980000   * iter 14, nobs 14000, eps 0.130800, edf 34.320000   * iter 15, nobs 15000, eps 0.293200, edf 27.300000   * iter 16, nobs 16000, eps 0.047700, edf 33.430000   * iter 17, nobs 17000, eps 0.038500, edf 28.680000   * iter 18, nobs 18000, eps 0.057900, edf 36.270000   * iter 19, nobs 19000, eps 0.060100, edf 32.030000   * iter 20, nobs 20000, eps 0.049900, edf 34.980000   * iter 21, nobs 21000, eps 0.080100, edf 35.780000   * iter 22, nobs 22000, eps 0.128500, edf 33.210000   * iter 23, nobs 23000, eps 0.071200, edf 39.120000   * iter 24, nobs 24000, eps 0.040200, edf 35.200000   * iter 25, nobs 25000, eps 0.080800, edf 30.180000   * iter 26, nobs 26000, eps 0.058700, edf 41.000000   * iter 27, nobs 27000, eps 0.043500, edf 33.150000   * iter 28, nobs 28000, eps 0.074900, edf 33.470000   * iter 29, nobs 29000, eps 0.035800, edf 35.390000   * iter 30, nobs 30000, eps 0.062900, edf 32.480000   * iter 31, nobs 31000, eps 0.059300, edf 40.860000   * iter 32, nobs 32000, eps 0.046400, edf 33.650000   * iter 33, nobs 33000, eps 0.073700, edf 37.790000   * iter 34, nobs 34000, eps 0.043800, edf 39.920000   * iter 35, nobs 35000, eps 0.046100, edf 34.220000   * iter 36, nobs 36000, eps 0.048000, edf 36.610000   * iter 37, nobs 37000, eps 0.046200, edf 37.960000   * iter 38, nobs 38000, eps 0.048500, edf 40.390000   * iter 39, nobs 39000, eps 0.089000, edf 35.650000   * iter 40, nobs 40000, eps 0.085900, edf 35.830000   * iter 41, nobs 41000, eps 0.049000, edf 29.700000   * iter 42, nobs 42000, eps 0.673600, edf 37.260000   * iter 43, nobs 43000, eps 0.069900, edf 39.720000   * iter 44, nobs 44000, eps 0.083900, edf 38.530000   * iter 45, nobs 45000, eps 0.059800, edf 41.980000   * iter 46, nobs 46000, eps 0.117200, edf 38.270000   * iter 47, nobs 47000, eps 0.073500, edf 33.470000   * iter 48, nobs 48000, eps 0.107500, edf 39.470000   * iter 49, nobs 49000, eps 0.050300, edf 35.130000   * iter 50, nobs 50000, eps 0.064800, edf 30.630000
#> 
#> elapsed time:  1.68min

## Plot coefficient paths for x3 in mu.
## pathplot(b2, name = "mu.s.s(x3).b")

## (3) Boosting type flavor, only update model term with
##     the largest contribution in the "out-of-sample"
##     log-likelihood. In this case, if edf = 0 during
##     runtime of the algorithm, no model has an additional
##     contribution and the algorithm converges. This
##     behavior is controlled by argument eps_loglik, the
##     higher eps_loglik, the more restrictive is the
##     updating step.

## Initialize intercepts.
set.seed(0)
b0 <- bamlss(num ~ 1, data = d, sampler = FALSE, optimizer = opt_bbfitp,
  batch_ids = batch_ids, slice = TRUE)
#> Error in eval(predvars, data, env): object 'num' not found

## Compute starting values, remove the first
## 10 iterates and compute the mean of the
## remaining iterates.
start <- coef(b0, FUN = mean, burnin = 10)
#> Error in coef(b0, FUN = mean, burnin = 10): object 'b0' not found

## Start boosting, only update if change in
## "out-of-sample" log-likelihood is 0.1
## eps_loglik = 0.001.
set.seed(3)
b3 <- bamlss(f, data = d, sampler = FALSE, optimizer = opt_bbfit,
    batch_ids = batch_ids, nu = 0.1, aic = TRUE, eps_loglik = 0.001,
    select = TRUE, always = FALSE, start = start)
#> Error in start[paste0(i, ".p.", colnames(x[[i]]$model.matrix))]: object of type 'closure' is not subsettable

## Plot log-likelihood contributions.
## contribplot(b3)
## In this case, the algorithm did not converge,
## we need more iterations/batches.

## Note, prediction uses last iterate.
p3 <- predict(b3, model = "mu")
#> Error in predict(b3, model = "mu"): object 'b3' not found

## (4) Use slice sampling under the "out-of-sample"
##     log likelihood for estimation of smoothing
##     variances. In this case model terms are always
##     updated ad parameter paths behave like a MCMC
##     chain. Therefore, use opt_bbfitp(), which stores
##     parameter paths as "mcmc" objects and we can
##     inspect using traceplots. Note nu = 1 if
##     slice = TRUE.
set.seed(4)
b4 <- bamlss(f, data = d, sampler = FALSE, optimizer = opt_bbfitp,
  batch_ids = batch_ids, aic = TRUE, slice = TRUE)
#>   .. df mu term s(x1) 4.000001 
#>   .. df mu term s(x2) 4.000001 
#>   .. df mu term s(x3) 4.000001 
#>   .. df mu term te(lon,lat) 10.00003 
#>   .. df sigma term s(x1) 4.000001 
#>   .. df sigma term s(x2) 4.000001 
#>   .. df sigma term s(x3) 4.000001 
#>   .. df sigma term te(lon,lat) 9.999995 
#>    * iter 1, nobs 1000, edf 39.438400   * iter 2, nobs 2000, eps 1.215400, edf 40.240000   * iter 3, nobs 3000, eps 0.621800, edf 46.580000   * iter 4, nobs 4000, eps 0.782600, edf 43.900000   * iter 5, nobs 5000, eps 1.157900, edf 42.880000   * iter 6, nobs 6000, eps 0.976200, edf 50.270000   * iter 7, nobs 7000, eps 2.127300, edf 50.340000   * iter 8, nobs 8000, eps 0.813100, edf 43.890000   * iter 9, nobs 9000, eps 0.829300, edf 42.660000   * iter 10, nobs 10000, eps 2.939000, edf 44.950000   * iter 11, nobs 11000, eps 0.587400, edf 35.870000   * iter 12, nobs 12000, eps 0.844700, edf 36.670000   * iter 13, nobs 13000, eps 2.006000, edf 39.610000   * iter 14, nobs 14000, eps 0.773200, edf 40.880000   * iter 15, nobs 15000, eps 4.226000, edf 41.830000   * iter 16, nobs 16000, eps 0.826100, edf 41.360000   * iter 17, nobs 17000, eps 2.108000, edf 37.940000   * iter 18, nobs 18000, eps 0.772200, edf 34.310000   * iter 19, nobs 19000, eps 0.669000, edf 37.680000   * iter 20, nobs 20000, eps 1.008600, edf 38.590000   * iter 21, nobs 21000, eps 0.539600, edf 35.150000   * iter 22, nobs 22000, eps 0.495500, edf 37.080000   * iter 23, nobs 23000, eps 0.633700, edf 34.660000   * iter 24, nobs 24000, eps 0.710100, edf 33.240000   * iter 25, nobs 25000, eps 0.827700, edf 32.960000   * iter 26, nobs 26000, eps 4.834700, edf 37.190000   * iter 27, nobs 27000, eps 0.973000, edf 32.990000   * iter 28, nobs 28000, eps 0.584500, edf 37.530000   * iter 29, nobs 29000, eps 1.213700, edf 43.680000   * iter 30, nobs 30000, eps 1.283300, edf 41.930000   * iter 31, nobs 31000, eps 0.831900, edf 40.040000   * iter 32, nobs 32000, eps 0.653000, edf 36.250000   * iter 33, nobs 33000, eps 1.376500, edf 38.770000   * iter 34, nobs 34000, eps 0.943700, edf 36.690000   * iter 35, nobs 35000, eps 0.602200, edf 32.510000   * iter 36, nobs 36000, eps 2.359600, edf 38.650000   * iter 37, nobs 37000, eps 1.831800, edf 36.300000   * iter 38, nobs 38000, eps 0.792300, edf 37.060000   * iter 39, nobs 39000, eps 0.568300, edf 35.730000   * iter 40, nobs 40000, eps 0.824400, edf 40.040000   * iter 41, nobs 41000, eps 0.843700, edf 43.720000   * iter 42, nobs 42000, eps 0.834300, edf 42.680000   * iter 43, nobs 43000, eps 0.577900, edf 40.800000   * iter 44, nobs 44000, eps 1.367300, edf 41.090000   * iter 45, nobs 45000, eps 1.117200, edf 33.470000   * iter 46, nobs 46000, eps 0.851700, edf 32.420000   * iter 47, nobs 47000, eps 0.977300, edf 45.490000   * iter 48, nobs 48000, eps 0.848000, edf 41.960000   * iter 49, nobs 49000, eps 1.012000, edf 39.750000   * iter 50, nobs 50000, eps 0.815700, edf 40.210000
#> 
#> elapsed time: 18.73sec

## plot(b4)

## Plot parameter updates/samples.
## plot(b4, which = "samples")

## Predict with burnin and compute mean
## prediction of the last 20 iterates.
p4 <- predict(b4, model = "mu", burnin = 30, FUN = mean)