Cox Model Posterior Mode Estimation

This function computes posterior mode estimates of the parameters of a flexible Cox model with structured additive predictors using a Newton-Raphson algorithm. Integrals are solved numerically. Moreover, optimum smoothing variances are computed using a stepwise optimization, see also the details section of function bfit.

cox.mode(x, y, start, weights, offset,
  criterion = c("AICc", "BIC", "AIC"),
  nu = 0.1, update.nu = TRUE,
  eps = .Machine$double.eps^0.25, maxit = 400,
  verbose = TRUE, digits = 4, ...)

Arguments

x
The x list, as returned from function bamlss.frame and transformed by function surv.transform, holding all model matrices and other information that is used for fitting the model.

y
The model response, as returned from function bamlss.frame.

start
A named numeric vector containing possible starting values, the names are based on function parameters.

weights
Prior weights on the data, as returned from function bamlss.frame.

offset
Can be used to supply model offsets for use in fitting, returned from function bamlss.frame.

criterion
Set the information criterion that should be used, e.g., for smoothing variance selection. Options are the corrected AIC "AICc", the "BIC" and "AIC".

nu
Calibrates the step length of parameter updates of one Newton-Raphson update.

update.nu
Should the updating step length be optimized in each iteration of the backfitting algorithm.

eps
The relative convergence tolerance of the backfitting algorithm.

maxit
The maximum number of iterations for the backfitting algorithm

verbose
Print information during runtime of the algorithm.

digits
Set the digits for printing when verbose = TRUE.

…
Currently not used.

Value

A list containing the following objects:

fitted.values

A named list of the fitted values of the modeled parameters of the selected distribution.

parameters

The estimated set regression coefficients and smoothing variances.

edf

The equivalent degrees of freedom used to fit the model.

logLik

The value of the log-likelihood.

logPost

The value of the log-posterior.

hessian

The Hessian matrix evaluated at the posterior mode.

converged

Logical, indicating convergence of the backfitting algorithm.

time

The runtime of the algorithm.

References

Umlauf N, Klein N, Zeileis A (2016). Bayesian Additive Models for Location Scale and Shape (and Beyond). (to appear)

See also

Examples

## Please see the examples of function cox.mcmc()!