HOMSTART Precipitation Data

This function downloads and compiles the HOMSTART-project data set. The data is downloaded from the Zentralanstalt fuer Meteorologie und Geodynamik (ZAMG, http://www.zamg.ac.at) and funded by the Austrian Climate Research Programme (ACRP) and is free for research purposes.

homstart_data(dir = NULL, load = TRUE, tdir = NULL)

Arguments

dir
The directory where the homstart.rda file should be stored.

load
Should the homstart data be loaded?

tdir
An optional temporary directory where all downloaded files are processed.

Value

A data frame containing the following variables:

raw

The daily precipitation observations.

cens

Precipitation observations censored at 0.

bin

Factor with levels "yes" or "no" indicating precipitation.

cat

Factor with levels "none", "low", "medium" and "high" indicating the amount of precipitation.

trend

A numeric time trend

month

Month of of the observation.

year

Year of the observation.

day

Day of the year.

lon

The longitude coordinate of the corresponding meteorological station.

lat

The latitude coordinate of the corresponding meteorological station.

id

Factor, meteorological station identifier.

cos1, cos2, sin1, sin2

Transformed time trend for harmonic regression.

weekend

Factor, with levels "yes" and "no" indication if the observation was measured on a weekend.

elevation

Numeric, the elevation of the meteorological station.

References

Nemec J, Gruber C, Chimani B, Auer I (2012). Trends in extreme temperature indices in Austria based on a new homogenised dataset. International Journal of Climatology. DOI 10.1002/joc.3532.

Nemec J, Chimani B, Gruber C, Auer I (2011). Ein neuer Datensatz homogenisierter Tagesdaten. OEGM Bulletin, 1, 19--20. http://www.meteorologie.at/docs/OEGM_bulletin_2011_1.pdf

Umlauf N, Mayr G, Messner J, Zeileis A (2012). Why does it always rain on me? A spatio-temporal analysis of precipitation in Austria. Austrian Journal of Statistics, 41(1), 81--92. http://www.stat.tugraz.at/AJS/ausg121/121Zeileis.pdf

Examples

# NOT RUN {
homstart_data(load = TRUE)
head(homstart)
# }